Moved Dash HTML out of app.py to a separate file for better legibility
